Rename RequestsReceived folder for consistency
